Add ClimateTargetPointExtender interface and update ClimateSettings to use Long for climate parameters

+++ common/src/main/kotlin/terrasect/definition/ClimateSettings.kt
@@ -1,14 +1,14 @@
 package terrasect.definition
 
-class ClimateRange(val min: Float, val max: Float) {
+class ClimateRange(val min: Long, val max: Long) {
   companion object {
-    fun range(min: Float, max: Float): ClimateRange {
+    fun range(min: Long, max: Long): ClimateRange {
       val lo = kotlin.math.min(min, max)
       val hi = kotlin.math.max(min, max)
       return ClimateRange(lo, hi)
     }
 
-    fun exact(value: Float) = ClimateRange(value, value)
+    fun exact(value: Long) = ClimateRange(value, value)
   }
 
   fun hasVariation() = min != max
@@ -19,14 +19,14 @@ class ClimateRange(val min: Float, val max: Float) {
 }
 
 data class ClimateSettings(
-  val temperature: ClimateRange? = null,
-  val humidity: ClimateRange? = null,
-  val continentalness: ClimateRange? = null,
-  val erosion: ClimateRange? = null,
-  val depth: ClimateRange? = null,
-  val weirdness: ClimateRange? = null,
-  val precipitation: String? = null,
-  val climatePreset: String? = null,
+    val temperature: ClimateRange? = null,
+    val humidity: ClimateRange? = null,
+    val continentalness: ClimateRange? = null,
+    val erosion: ClimateRange? = null,
+    val depth: ClimateRange? = null,
+    val weirdness: ClimateRange? = null,
+    val precipitation: String? = null,
+    val climatePreset: String? = null,
 ) {
 
   companion object {
@@ -43,33 +43,33 @@ data class ClimateSettings(
     private var precipitation: String? = null
     private var climatePreset: String? = null
 
-    fun temperature(min: Float, max: Float) = apply {
+    fun temperature(min: Long, max: Long) = apply {
       this.temperature = ClimateRange.range(min, max)
     }
 
-    fun temperature(value: Float) = apply { this.temperature = ClimateRange.exact(value) }
+    fun temperature(value: Long) = apply { this.temperature = ClimateRange.exact(value) }
 
-    fun humidity(min: Float, max: Float) = apply { this.humidity = ClimateRange.range(min, max) }
+    fun humidity(min: Long, max: Long) = apply { this.humidity = ClimateRange.range(min, max) }
 
-    fun humidity(value: Float) = apply { this.humidity = ClimateRange.exact(value) }
+    fun humidity(value: Long) = apply { this.humidity = ClimateRange.exact(value) }
 
-    fun continentalness(min: Float, max: Float) = apply {
+    fun continentalness(min: Long, max: Long) = apply {
       this.continentalness = ClimateRange.range(min, max)
     }
 
-    fun continentalness(value: Float) = apply { this.continentalness = ClimateRange.exact(value) }
+    fun continentalness(value: Long) = apply { this.continentalness = ClimateRange.exact(value) }
 
-    fun erosion(min: Float, max: Float) = apply { this.erosion = ClimateRange.range(min, max) }
+    fun erosion(min: Long, max: Long) = apply { this.erosion = ClimateRange.range(min, max) }
 
-    fun erosion(value: Float) = apply { this.erosion = ClimateRange.exact(value) }
+    fun erosion(value: Long) = apply { this.erosion = ClimateRange.exact(value) }
 
-    fun depth(min: Float, max: Float) = apply { this.depth = ClimateRange.range(min, max) }
+    fun depth(min: Long, max: Long) = apply { this.depth = ClimateRange.range(min, max) }
 
-    fun depth(value: Float) = apply { this.depth = ClimateRange.exact(value) }
+    fun depth(value: Long) = apply { this.depth = ClimateRange.exact(value) }
 
-    fun weirdness(min: Float, max: Float) = apply { this.weirdness = ClimateRange.range(min, max) }
+    fun weirdness(min: Long, max: Long) = apply { this.weirdness = ClimateRange.range(min, max) }
 
-    fun weirdness(value: Float) = apply { this.weirdness = ClimateRange.exact(value) }
+    fun weirdness(value: Long) = apply { this.weirdness = ClimateRange.exact(value) }
 
     fun precipitation(precipitation: String) = apply { this.precipitation = precipitation }
 
@@ -87,15 +87,15 @@ data class ClimateSettings(
     }
 
     fun build(): ClimateSettings =
-      ClimateSettings(
-        temperature = temperature,
-        humidity = humidity,
-        continentalness = continentalness,
-        erosion = erosion,
-        depth = depth,
-        weirdness = weirdness,
-        precipitation = precipitation,
-        climatePreset = climatePreset,
-      )
+        ClimateSettings(
+            temperature = temperature,
+            humidity = humidity,
+            continentalness = continentalness,
+            erosion = erosion,
+            depth = depth,
+            weirdness = weirdness,
+            precipitation = precipitation,
+            climatePreset = climatePreset,
+        )
   }
 }
